Move bounds to same hierarchy level as algebra
