+++ src/main/java/de/uni_kl/cs/discodnc/bounds/Backlog.java
@@ -27,7 +27,7 @@
  *
  */
 
-package de.uni_kl.cs.discodnc.nc.bounds;
+package de.uni_kl.cs.discodnc.bounds;
 
 import de.uni_kl.cs.discodnc.Calculator;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
