+++ src/main/java/de/uni_kl/cs/discodnc/bounds/Output.java
@@ -26,7 +26,7 @@
  *
  */
 
-package de.uni_kl.cs.discodnc.nc.bounds;
+package de.uni_kl.cs.discodnc.bounds;
 
 import de.uni_kl.cs.discodnc.Calculator;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
